+++ src/main/resources/db/migration/V1__Initial_Setup.sql
@@ -7,7 +7,7 @@ CREATE TABLE guild
 
 CREATE TABLE embed_type
 (
-    id      INT NOT NULL,
+    id      BIGSERIAL NOT NULL,
     name    VARCHAR(150) NOT NULL,
     structure TEXT NOT NULL,
     PRIMARY KEY (id)
@@ -23,7 +23,7 @@ CREATE TABLE event
     dest_channel VARCHAR(30) NOT NULL,
     member_size  VARCHAR(5)  NOT NULL,
     guild_id     VARCHAR(30) NOT NULL,
-    embed_type   INT         NOT NULL,
+    embed_type   BIGINT         NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (guild_id) REFERENCES guild(id) ON DELETE CASCADE,
     FOREIGN KEY (embed_type) REFERENCES embed_type(id) ON DELETE CASCADE
